requirements.txt
auto-sklearn==0.6.0
autogluon==0.0.6
autoPyTorch==0.0.2
autokeras==1.0.2
h2o==3.30.0.2
keras-tuner==1.0.1
numpy==1.18.3
pandas==1.0.3
scikit-learn==0.22.1
TPOT==0.11.1
